fix(Scripts/Karazhan): use size_t within loop in boss_shade_of_aran:Reset()

+++ src/server/scripts/EasternKingdoms/Karazhan/boss_shade_of_aran.cpp
@@ -124,7 +124,7 @@ struct boss_shade_of_aran : public BossAI
         _drinking = false;
         _hasDrunk = false;
 
-        for (int i = 0; i < immuneSpells.size(); i++)
+        for (size_t i = 0; i < immuneSpells.size(); i++)
             me->ApplySpellImmune(0, IMMUNITY_ID, immuneSpells[i], true);
 
         if (GameObject* libraryDoor = instance->instance->GetGameObject(instance->GetGuidData(DATA_GO_LIBRARY_DOOR)))
